fix(dap): make debug tools available in every phase and mode
DAP tools were being filtered out during ferment planning and adhoc
planning modes because the tool-profile system (tool-catalog.ts) only
includes tools explicitly listed in SHARED_CORE_TOOLS. DAP tools
weren't in the catalog, so getToolsForProfile() excluded them from
the active tool list in planning profiles.

Fix: add all 14 DAP tool names to SHARED_CORE_TOOLS so they're included
in every profile (idle, planning-adhoc, planning-ferment,
implementation-ferment). This makes debug tools always available
regardless of phase or mode — the agent can inspect runtime state
at any time, not just during build/review.

Also removes the phase-based visibility filtering from dap.ts:
- Remove DAP_VISIBLE_PHASES constant
- Remove DAP_TOOL_NAMES constant (no longer needed for visibility toggling)
- Remove the tool_call handler that polled getCurrentPhase and toggled
  visibility
- Remove createToolVisibility and getCurrentPhase imports

Deletes src/extensions/dap/phase-visibility.test.ts (obsolete).
Updates dap-entry.test.ts to assert tools are always visible across
all phases instead of testing hide/show transitions.
