+++ src/shared/planning/tool-catalog.ts
@@ -109,6 +109,22 @@ export const SHARED_CORE_TOOLS: ToolEntry[] = [
 	// servers are configured, and required during planning so the model can
 	// search/describe/call read-only MCP tools (e.g. Atlassian Jira).
 	{ name: "mcp", modes: ["shared"] },
+	// DAP debugger tools — always available in every mode/profile so the agent
+	// can inspect runtime state at any time. Registered by the dap extension.
+	{ name: "debug_launch", modes: ["shared"] },
+	{ name: "debug_set_breakpoint", modes: ["shared"] },
+	{ name: "debug_continue", modes: ["shared"] },
+	{ name: "debug_locals", modes: ["shared"] },
+	{ name: "debug_eval", modes: ["shared"] },
+	{ name: "debug_backtrace", modes: ["shared"] },
+	{ name: "debug_terminate", modes: ["shared"] },
+	{ name: "step_in", modes: ["shared"] },
+	{ name: "step_over", modes: ["shared"] },
+	{ name: "step_out", modes: ["shared"] },
+	{ name: "debug_state_at", modes: ["shared"] },
+	{ name: "debug_last_error", modes: ["shared"] },
+	{ name: "debug_trace_calls", modes: ["shared"] },
+	{ name: "debug_watch_change", modes: ["shared"] },
 	// Todo lifecycle tools — must mirror TODO_TOOL_NAMES in src/extensions/todos/tool.ts
 	// These are general-purpose session tools used in all modes (adhoc chat,
 	// ferment planning, and ferment implementation). The system prompt,
